MidiCCButton構造体内の変数の型誤り修正
VMCProtocolの仕様上、MIDI CC Button Inputの第二引数はint型ですが、MidiCCButton構造体内で宣言されている第二引数用の変数がfloat型で宣言されていますので、修正しました。

+++ EVMC4U/class/EVMC4Ustructs.cs
@@ -74,7 +74,7 @@ namespace EVMC4U {
     public struct MidiCCButton
     {
         public int knob;
-        public float active;
+        public int active;
     }
 
     public enum CalibrationState
